+++ Pod/SyneriseSDK.framework/Headers/SNRClientAuthenticationStatus.h
@@ -16,6 +16,7 @@ typedef NS_ENUM(NSUInteger, SNRClientAuthenticationStatus) {
     SNRClientAuthenticationStatusSuccess,
     SNRClientAuthenticationStatusUnauthorized,
     SNRClientAuthenticationStatusActivationRequired,
+    SNRClientAuthenticationStatusPinActivationRequired,
     SNRClientAuthenticationStatusRegistrationRequired,
     SNRClientAuthenticationStatusApprovalRequired,
     SNRClientAuthenticationStatusTermsAcceptanceRequired,