+++ Pod/SyneriseSDK.framework/Headers/SNRErrorCode.h
@@ -28,7 +28,9 @@ typedef NS_ENUM(NSInteger, SNRErrorCode) {
     SNRErrorCodeInvalidResponseApiError = -106,
     SNRErrorCodeClientSessionUnauthorizedApiError = 3200,
     SNRErrorCodeClientSessionExpiredApiError = 3210,
-    SNRErrorCodeJWTValidationFailedApiError = 3300
+    SNRErrorCodeJWTValidationFailedApiError = 3300,
+    SNRErrorCodeJWTProcessingFailedApiError = 3310,
+    SNRErrorCodeMissingPublicKeyApiError = 3320,
 };
 
 NS_ASSUME_NONNULL_END